+++ pandas/api/__init__.py
@@ -1,5 +1,6 @@
 """ public toolkit API """
 from pandas.api import (  # noqa:F401
+    exchange,
     extensions,
     indexers,
     types,
