+++ pandas/tests/exchange/test_protocol.py
@@ -1,3 +1,6 @@
+"""
+A verbatim copy (vendored) of the spec tests from https://github.com/data-apis/dataframe-api
+"""
 import pytest
 import math
 import ctypes
